[fix] 로그인 하지 않은채로 운동 계획
#141
