+++ src/screens/ProfileScreen/ProfileScreen.test.tsx
@@ -1,5 +1,5 @@
-import { navigationMock, navigationNavigateMock } from '@mocks/navigationMocks';
-import { act, fireEvent, render } from '@testing-library/react-native';
+import { navigationMock } from '@mocks/navigationMocks';
+import { render } from '@testing-library/react-native';
 import React from 'react';
 
 import { wrapper } from '@tests/functions';
@@ -23,15 +23,5 @@ describe('ProfileScreen 컴포넌트', () => {
     const { toJSON } = rendered();
 
     expect(toJSON()).toMatchSnapshot();
-  });
-
-  it('로그인 하지 않은 사용자가 로그인 하러가기 버튼을 누를 수 있다', async () => {
-    const { getByText } = rendered();
-
-    await act(async () => fireEvent.press(getByText('로그인 하러가기')));
-
-    expect(navigationNavigateMock).toHaveBeenCalledWith('Auth', {
-      screen: 'Login',
-    });
   });
 });
